# Use a space trim implementation that's optimized for ASCII