Before:

```
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    5000	    391574 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	^[      50	  36527847 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	3.888s
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    5000	    429596 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	      50	  36930323 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	4.102s
```

After:

```
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    5000	    325336 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	      50	  31389034 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	3.287s
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    5000	    330157 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	      50	  30757109 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	3.283s
```