@alex 2019-08-25
FWIW, it looks like Go 1.13 has basically this same optimization, so if you want to drop this and just tell users to upgrade to 1.13 which will be out pretty soon, that's also fine: https://github.com/golang/go/commit/4b4f222a0dd8765e5b493d458fa352ea22045575