@ianlopshire 2019-08-26
@alex,

I'm torn on this, but I think I'd rather drop it and wait for Go 1.13. Closing for now, but let me know if you'd like me to reconsider.