@alex 2019-08-26
I think Go 1.13 is supposed to be out pretty soon, so this is fine with me -- I didn't realize upstream had done the same thing at the time I wrote this.